// backend/src/models/commonStructure.zod.gen.ts
export const identifierTypeSchema = z.union([
  z.literal('Audiovisual'),
  z.literal('Book'),
  z.literal('BookChapter'),
  z.literal('Collection'),
  z.literal('ComputationalNotebook'),
  z.literal('ConferencePaper'),
  z.literal('ConferenceProceeding'),
  z.literal('DataPaper'),
  z.literal('Dataset'),
  z.literal('Dissertation'),
  z.literal('Event'),
  z.literal('Image'),
  z.literal('InteractiveResource'),
  z.literal('Journal'),
  z.literal('JournalArticle'),
  z.literal('Model'),
  z.literal('OutputManagementPlan'),
  z.literal('PeerReview'),
  z.literal('PhysicalObject'),
  z.literal('Preprint'),
  z.literal('Report'),
  z.literal('Service'),
  z.literal('Software'),
  z.literal('Sound'),
  z.literal('Standard'),
  z.literal('Text'),
  z.literal('Workflow'),
  z.literal('Other'),
  z.literal('ARK'),
  z.literal('arXiv'),
  z.literal('bibcode'),
  z.literal('DOI'),
  z.literal('EAN13'),
  z.literal('EISSN'),
  z.literal('Handle'),
  z.literal('ISBN'),
  z.literal('ISSN'),
  z.literal('ISTC'),
  z.literal('LISSN'),
  z.literal('LSID'),
  z.literal('ORCID'),
  z.literal('PMID'),
  z.literal('PURL'),
  z.literal('UPC'),
  z.literal('URL'),
  z.literal('URN'),
  z.literal('w3id'),
]);

export const alternateIdentifierSchema = z.object({
  alternateID: z.string(),
  alternateIDType: identifierTypeSchema,
});

export const titleSchema = z.object({
  titleLanguage: z.union([z.string(), z.undefined()]).optional(),
  titleText: z.string(),
});

export const contactSchema = z.object({
  contactName: z.string().optional(),
  contactEmail: z.string().optional(),
});

export const descriptionSchema = z.object({
  descriptionText: z.string(),
  descriptionType: z.union([z.string(), z.undefined()]).optional(),
});

export const keywordsSchema = z.object({
  keywordLabel: z.string(),
  keywordURI: z.string().optional(),
});

export const temporalCoverageSchema = z.object({
  startDate: z.string().optional(),
  endDate: z.string().optional(),
});

export const licenseSchema = z.object({
  licenseCode: z.string().optional(),
  licenseURI: z.string().optional(),
});

export const fileSchema = z.object({
  name: z.string().optional(),
  sourceUrl: z.string().optional(),
  format: z.string().optional(),
  md5: z.string().optional(),
  size: z.string().optional(),
  sizeMeasureType: z.union([z.literal('GB'), z.literal('MB'), z.literal('kB'), z.literal('B')]).optional(),
});

export const projectSchema = z.object({
  projectName: z.string().optional(),
  projectID: z.string().optional(),
});

export const siteReferenceSchema = z.object({
  siteID: z.string().optional(),
  siteName: z.string().optional(),
});

export const additionalMetadataSchema = z.object({
  name: z.string(),
  value: z.string(),
});

export const habitatReferenceSchema = z.object({
  soHabitatCode: z.union([
    z.literal('A'),
    z.literal('B'),
    z.literal('C1'),
    z.literal('C2'),
    z.literal('D'),
    z.literal('E'),
    z.literal('F'),
    z.literal('G'),
    z.literal('H'),
    z.literal('I'),
  ]),
  soHabitatURI: z.string().optional(),
});

export const dataLevelSchema = z.object({
  dataLevelCode: z.string().optional(),
  dataLevelURI: z.string().optional(),
});

export const externalSourceSchema = z.object({
  externalSourceName: z.string().optional(),
  externalSourceURI: z.string().optional(),
  externalSourceInfo: z.string().optional(),
});

export const doiSchema = z.object({
  identifier: z.string(),
  provider: z.string(),
});

export const pidSchema = z.object({
  doi: doiSchema,
});
// ...
export const relatedIdentifierSchema = z.object({
  relatedID: z.string(),
  relatedIDType: identifierTypeSchema,
  relatedResourceType: identifierTypeSchema,
  relationType: relationSchema,
});

export const organizationSchema = z.object({
  organizationName: z.string().optional(),
  organizationEmail: z.string().optional(),
  organizationIDs: z.array(entityIdentifierSchema).optional(),
});

export const affiliationSchema = z.object({
  entityName: z.string().optional(),
  entityID: entityIdentifierSchema.optional(),
});

export const responsibleOrganizationsSchema = z.object({
  organizationName: z.string().optional(),
  organizationEmail: z.string().optional(),
  organizationIDs: z.array(entityIdentifierSchema).optional(),
});
// ...
export const contributorSchema = z.object({
  contributorFamilyName: z.string().optional(),
  contributorGivenName: z.string().optional(),
  contributorEmail: z.string().optional(),
  contributorAffiliation: affiliationSchema.optional(),
  contributorIDs: z.array(entityIdentifierSchema).optional(),
  contributorType: contributorTypeSchema.optional(),
});
// ...
export const datasetTypeSchema = z.object({
  datasetTypeCode: datasetTypeCodeSchema,
  datasetTypeURI: z.string().optional(),
});

export const creatorSchema = z.object({
  creatorFamilyName: z.string().optional(),
  creatorGivenName: z.string().optional(),
  creatorEmail: z.string().optional(),
  creatorAffiliation: affiliationSchema.optional(),
  creatorIDs: z.array(entityIdentifierSchema).optional(),
});

export const geolocationSchema = z.object({
  observationLocation: observationLocationSchema.optional(),
  geographicDescription: z.string().optional(),
  boundingBox: boundingBoxSchema.optional(),
  boundingPolygon: z
    .array(
      z.object({
        points: z.array(coordinatesSchema),
        inPolygonPoint: coordinatesSchema,
      }),
    )
    .optional(),
  lineString: z.array(coordinatesSchema).optional(),
  point: coordinatesSchema.optional(),
});

export const methodSchema = z.object({
  methodID: z.string().optional(),
  steps: z.array(stepsSchema).optional(),
  sampling: samplingSchema.optional(),
  qualityControlDescription: z.string().optional(),
  instrumentationDescription: z.string().optional(),
});

export const commonDatasetMetadataSchema = z.object({
  assetType: identifierTypeSchema,
  datasetType: datasetTypeSchema.optional(),
  alternateIdentifiers: z.array(alternateIdentifierSchema).optional(),
  titles: z.array(titleSchema).optional(),
  creators: z.array(creatorSchema).optional(),
  contactPoints: z.array(contactSchema).optional(),
  descriptions: z.array(descriptionSchema).optional(),
  keywords: z.array(keywordsSchema).optional(),
  temporalCoverages: z.array(temporalCoverageSchema).optional(),
  geoLocations: z.array(geolocationSchema).optional(),
  licenses: z.array(licenseSchema).optional(),
  files: z.array(fileSchema).optional(),
  responsibleOrganizations: z.array(responsibleOrganizationsSchema).optional(),
  contributors: z.array(contributorSchema).optional(),
  publicationDate: z.string().optional(),
  taxonomicCoverages: z.array(z.string()).optional(),
  methods: z.array(methodSchema).optional(),
  language: z.string().optional(),
  projects: z.array(projectSchema).optional(),
  siteReferences: z.array(siteReferenceSchema).optional(),
  additionalMetadata: z.array(additionalMetadataSchema).optional(),
  habitatReferences: z.array(habitatReferenceSchema).optional(),
  relatedIdentifiers: z.array(relatedIdentifierSchema).optional(),
  dataLevel: dataLevelSchema.optional(),
  externalSourceInformation: externalSourceSchema,
});

export const commonDatasetSchema = z.object({
  pids: pidSchema.optional(),
  metadata: commonDatasetMetadataSchema,
});